+++ src/mesh_old/util/tex_coords.rs
