+++ src/platform_io/mod.rs
@@ -28,7 +28,6 @@ impl Variant {
 
 // ---- Export ----------------------------------------------------------------
 mod export;
-use std::result;
 
 pub use export::export;
 pub use export::export_to_glb;
@@ -43,14 +42,12 @@ pub use import::import;
 // ---- Result & Errors ----------------------------------------------------------
 
 /// Result type for convenience.
-pub type Result<T> = result::Result<T, Error>;
+pub type Result<T> = std::result::Result<T, Error>;
 
 /// Represents a runtime error.
 #[derive(Debug)]
 pub enum Error {
     /// Base 64 decoding error.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     Base64(base64::DecodeError),
 
     /// GLB parsing error.
@@ -59,8 +56,6 @@ pub enum Error {
     InvalidExtension,
 
     /// Buffer length does not match expected length.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     BufferLength {
         /// The index of the offending buffer.
         buffer: usize,
@@ -79,33 +74,21 @@ pub enum Error {
     Io(std::io::Error),
 
     /// Image decoding error.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     Image(image_crate::ImageError),
 
     /// The `BIN` chunk of binary glTF is referenced but does not exist.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     MissingBlob,
 
     /// An external file is referenced in a slice only import without path
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     ExternalReferenceInSliceImport,
 
     /// Unsupported image encoding.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     UnsupportedImageEncoding,
 
     /// Unsupported image format.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     UnsupportedImageFormat(image_crate::DynamicImage),
 
     /// Unsupported URI scheme.
-    #[cfg(feature = "import")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "import")))]
     UnsupportedScheme,
 
     /// glTF validation error.
