+++ src/properties/animation.rs
@@ -1,0 +1,1 @@
+pub struct Animation(gltf_json::Animation);
