+++ src/properties/image.rs
@@ -1,0 +1,1 @@
+pub struct Image(gltf_json::Image);
