+++ src/properties/material.rs
@@ -1,0 +1,1 @@
+pub struct Material(gltf_json::Material);
