+++ src/properties/mesh.rs
@@ -1,0 +1,1 @@
+pub struct Mesh(gltf_json::Mesh);
