+++ src/properties/node.rs
@@ -1,0 +1,1 @@
+pub struct Node(gltf_json::Node);
