+++ src/properties/scene.rs
@@ -1,0 +1,1 @@
+pub struct Scene(gltf_json::Scene);
