+++ src/properties/skin.rs
@@ -1,0 +1,1 @@
+pub struct Skin(gltf_json::Skin);
