+++ src/properties/texture.rs
@@ -1,0 +1,1 @@
+pub struct Texture(gltf_json::Texture);
